feat: prediction market engine complete
Finishes Component 3 of 5: the prediction-market engine that prices
uncertainty for the rest of the Lace protocol.

This commit adds:

* lace-pm-compose -- the composability facade. Engine struct that
  holds markets, AMM state, and conditional-trigger registry. Exposes
  ProbabilityFeed (get_live_probability, get_resolved_outcome,
  create_conditional_trigger) and OracleResolver, the latter matching
  lace-conditions::OracleResolver exactly so a one-trait adapter
  bridges to the temporal VM.

* lace-pm-governance -- prediction-gated upgrade execution and
  parameter governance. Proposals only execute if their bound forecast
  market resolves YES AND the market accumulated enough volume to be
  considered binding (defends against thin-market manipulation).

* Workspace integration tests in lace-pm-governance/tests covering the
  full lifecycle: binary market with resolution and AMM trading;
  conditional cascade-voiding; upheld disputes overturning provisional
  resolutions; governance executing through the full pipeline; AMM
  position-commitment opacity; trigger callbacks firing into temporal
  VM-style sinks; manipulation resistance for thin markets; scalar and
  multi-outcome resolution paths.

* INTEGRATION.md documenting every cross-component interface, why the
  OracleResolver trait is duplicated rather than imported across
  workspaces, and the calibration window semantics for Veil Score
  feedback.

* Clippy pass clean across the workspace.

75 tests passing in total (6 types, 9 markets, 14 AMM, 12 oracle,
13 compose, 8 governance, 13 integration).

+++ protocol/prediction-markets/crates/lace-pm-markets/src/lib.rs
@@ -267,8 +267,8 @@ pub fn scalar_to_probability(lo: i128, hi: i128, s: i128) -> Probability {
     if s >= hi {
         return Probability::ONE;
     }
-    let num = (s - lo) as i128;
-    let den = (hi - lo) as i128;
+    let num = s - lo;
+    let den = hi - lo;
     // Convert to basis points with integer math to keep this
     // deterministic across architectures.
     let bps = (num.saturating_mul(10_000) / den) as u32;
